+++ tests/sitemap-index.test.ts
@@ -1,7 +1,17 @@
-import 'babel-polyfill';
-import { buildSitemapIndex, createSitemapsAndIndex } from '../index';
+import { createSitemapsAndIndex, SitemapStream } from '../index';
 import { tmpdir } from 'os';
-import { existsSync, unlinkSync } from 'fs';
+import { resolve } from 'path';
+import {
+  existsSync,
+  unlinkSync,
+  createWriteStream,
+  createReadStream,
+} from 'fs';
+import {
+  SitemapIndexStream,
+  SitemapAndIndexStream,
+} from '../lib/sitemap-index-stream';
+import { streamToPromise } from '../dist';
 /* eslint-env jest, jasmine */
 function removeFilesArray(files): void {
   if (files && files.length) {
@@ -15,7 +25,7 @@ function removeFilesArray(files): void {
 
 const xmlDef = '<?xml version="1.0" encoding="UTF-8"?>';
 describe('sitemapIndex', () => {
-  it('build sitemap index', () => {
+  it('build sitemap index', async () => {
     const expectedResult =
       xmlDef +
       '<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">' +
@@ -26,35 +36,16 @@ describe('sitemapIndex', () => {
       '<loc>https://test.com/s2.xml</loc>' +
       '</sitemap>' +
       '</sitemapindex>';
+    const smis = new SitemapIndexStream();
+    smis.write('https://test.com/s1.xml');
+    smis.write('https://test.com/s2.xml');
+    smis.end();
+    const result = await streamToPromise(smis);
 
-    const result = buildSitemapIndex({
-      urls: ['https://test.com/s1.xml', 'https://test.com/s2.xml'],
-    });
-
-    expect(result).toBe(expectedResult);
+    expect(result.toString()).toBe(expectedResult);
   });
 
-  it('build sitemap index with custom xmlNS', () => {
-    const expectedResult =
-      xmlDef +
-      '<sitemapindex xmlns="http://www.example.org/schemas/sitemap/0.9">' +
-      '<sitemap>' +
-      '<loc>https://test.com/s1.xml</loc>' +
-      '</sitemap>' +
-      '<sitemap>' +
-      '<loc>https://test.com/s2.xml</loc>' +
-      '</sitemap>' +
-      '</sitemapindex>';
-
-    const result = buildSitemapIndex({
-      urls: ['https://test.com/s1.xml', 'https://test.com/s2.xml'],
-      xmlNs: 'xmlns="http://www.example.org/schemas/sitemap/0.9"',
-    });
-
-    expect(result).toBe(expectedResult);
-  });
-
-  it('build sitemap index with lastmodISO', () => {
+  it('build sitemap index with lastmodISO', async () => {
     const expectedResult =
       xmlDef +
       '<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">' +
@@ -68,52 +59,26 @@ describe('sitemapIndex', () => {
       '</sitemap>' +
       '<sitemap>' +
       '<loc>https://test.com/s3.xml</loc>' +
-      '<lastmod>2019-07-01T00:00:00.000Z</lastmod>' +
       '</sitemap>' +
       '</sitemapindex>';
-
-    const result = buildSitemapIndex({
-      urls: [
-        {
-          url: 'https://test.com/s1.xml',
-          lastmod: '2018-11-26',
-        },
-        {
-          url: 'https://test.com/s2.xml',
-          lastmod: '2018-11-27',
-        },
-        {
-          url: 'https://test.com/s3.xml',
-        },
-      ],
-      xmlNs: 'xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"',
-      lastmod: '2019-07-01',
-    });
-
-    expect(result).toBe(expectedResult);
-  });
-
-  it('build sitemap index with lastmod', () => {
-    const expectedResult =
-      xmlDef +
-      '<sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">' +
-      '<sitemap>' +
-      '<loc>https://test.com/s1.xml</loc>' +
-      '<lastmod>2018-11-26T00:00:00.000Z</lastmod>' +
-      '</sitemap>' +
-      '</sitemapindex>';
-
-    const result = buildSitemapIndex({
-      urls: [
-        {
-          url: 'https://test.com/s1.xml',
-        },
-      ],
-      xmlNs: 'xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"',
+    const smis = new SitemapIndexStream();
+    smis.write({
+      url: 'https://test.com/s1.xml',
       lastmod: '2018-11-26',
     });
 
-    expect(result).toBe(expectedResult);
+    smis.write({
+      url: 'https://test.com/s2.xml',
+      lastmod: '2018-11-27',
+    });
+
+    smis.write({
+      url: 'https://test.com/s3.xml',
+    });
+    smis.end();
+    const result = await streamToPromise(smis);
+
+    expect(result.toString()).toBe(expectedResult);
   });
 
   it('simple sitemap index', async () => {
@@ -182,5 +147,63 @@ describe('sitemapIndex', () => {
     expectedFiles.forEach(function(expectedFile) {
       expect(existsSync(expectedFile)).toBe(true);
     });
+  });
+});
+
+describe('sitemapAndIndex', () => {
+  let targetFolder: string;
+
+  beforeEach(() => {
+    targetFolder = tmpdir();
+    removeFilesArray([
+      resolve(targetFolder, `./sitemap-0.xml`),
+      resolve(targetFolder, `./sitemap-1.xml`),
+      resolve(targetFolder, `./sitemap-2.xml`),
+      resolve(targetFolder, `./sitemap-3.xml`),
+    ]);
+  });
+
+  afterEach(() => {
+    removeFilesArray([
+      resolve(targetFolder, `./sitemap-0.xml`),
+      resolve(targetFolder, `./sitemap-1.xml`),
+      resolve(targetFolder, `./sitemap-2.xml`),
+      resolve(targetFolder, `./sitemap-3.xml`),
+    ]);
+  });
+
+  it('writes both a sitemap and index', async () => {
+    const baseURL = 'https://example.com/sub/';
+
+    const sms = new SitemapAndIndexStream({
+      limit: 1,
+      getSitemapStream: (i: number): [string, SitemapStream] => {
+        const sm = new SitemapStream();
+        const path = `./sitemap-${i}.xml`;
+
+        sm.pipe(createWriteStream(resolve(targetFolder, path)));
+        return [new URL(path, baseURL).toString(), sm];
+      },
+    });
+    sms.write('https://1.example.com/a');
+    sms.write('https://2.example.com/a');
+    sms.write('https://3.example.com/a');
+    sms.write('https://4.example.com/a');
+    sms.end();
+    const index = (await streamToPromise(sms)).toString();
+    expect(index).toContain(`${baseURL}sitemap-0`);
+    expect(index).toContain(`${baseURL}sitemap-1`);
+    expect(index).toContain(`${baseURL}sitemap-2`);
+    expect(index).toContain(`${baseURL}sitemap-3`);
+    expect(index).not.toContain(`${baseURL}sitemap-4`);
+    expect(existsSync(resolve(targetFolder, `./sitemap-0.xml`))).toBe(true);
+    expect(existsSync(resolve(targetFolder, `./sitemap-1.xml`))).toBe(true);
+    expect(existsSync(resolve(targetFolder, `./sitemap-2.xml`))).toBe(true);
+    expect(existsSync(resolve(targetFolder, `./sitemap-3.xml`))).toBe(true);
+    expect(existsSync(resolve(targetFolder, `./sitemap-4.xml`))).toBe(false);
+    const xml = await streamToPromise(
+      createReadStream(resolve(targetFolder, `./sitemap-0.xml`))
+    );
+    expect(xml.toString()).toContain('https://1.example.com/a');
   });
 });